Use settings API

+++ admin/views/render-reset.php
@@ -1,48 +1,34 @@
 <?php
 
 // If this file is called directly, abort.
-if (!defined('WPINC')) {
-    die;
+if ( ! defined( 'WPINC' ) ) {
+	die;
 }
 
 /**
  * Pick up options from the DB, check them against the whitelist
  */
 $options_inDB = $this->scan_options_table();
-$options = array_intersect($this->options_whitelist, $options_inDB);
+$options      = array_intersect( $this->options_whitelist, $options_inDB );
 
 /**
  * If this is the postback for reset
  */
-if (isset($_POST['reset']) && $_POST['reset'] === 'true' && check_admin_referer('deletefacebook_' . get_current_user_id())) {
+if ( isset( $_POST['reset'] ) && $_POST['reset'] === 'true' && check_admin_referer( 'deletefacebook_' . get_current_user_id() ) ) {
 
-    foreach ($this->options_whitelist as $option) {
-        delete_option($option);
+	if( $options ){
+		foreach ( $options as $option ) {
+			delete_option( $option );
+		}
     }
 
-    /**
-     * Pick up options from the DB, check them against the whitelist
-     */
-    $options_inDB = $this->scan_options_table();
-    $options = array_intersect($this->options_whitelist, $options_inDB);
+	/**
+	 * Pick up options from the DB, check them against the whitelist
+	 */
+	$options_inDB = $this->scan_options_table();
+	$options      = array_intersect( $this->options_whitelist, $options_inDB );
 
 }
-
-/**
- * If this is the postback for proxy
- */
-if (isset($_POST['proxy']) && $_POST['proxy'] === 'true' && check_admin_referer('proxyconnection_' . get_current_user_id())) {
-
-    /**
-     * Pick up options from the DB, check them against the whitelist
-     */
-    $set_proxy = ( isset( $_POST['proxyconnection'] ) ) ? intval( $_POST['proxyconnection'] ) : 0;
-    update_option('tfb4wc_set_proxy', $set_proxy);
-
-}else{
-
-}
-
 
 ?>
 
@@ -51,81 +37,90 @@ if (isset($_POST['proxy']) && $_POST['proxy'] === 'true' && check_admin_referer(
     <div id="icon-themes" class="icon32"></div>
     <h2>Troubleshoot Facebook for WooCommerce</h2>
 
-    <hr>
+	<?php $active_tab = ( isset( $_GET['tab'] ) ) ? sanitize_text_field( $_GET['tab'] ) : 'delete-options'; ?>
 
-    <h3>Delete all WooCommerce for Facebook options</h3>
-    <h4>This is only for when you want to completely reset your connection with Facebook, after all else has
-        failed.</h4>
+    <h2 class="nav-tab-wrapper">
+        <a href="?page=troubleshoot-facebook-for-woocommerce&tab=delete-options"
+           class="nav-tab <?php echo 'delete-options' === $active_tab ? 'nav-tab-active' : ''; ?>">
+            Delete options
+        </a>
 
-    <h4>You have <?php echo esc_html(count($options)); ?> Facebook for WooCommerce options stored in your database.</h4>
+        <a href="?page=troubleshoot-facebook-for-woocommerce&tab=connection-test"
+           class="nav-tab <?php echo 'connection-test' === $active_tab ? 'nav-tab-active' : ''; ?>">
+            Connection test
+        </a>
 
-    <?php
-    /**
-     * Only if Facebook for WooCommerce is not active, and there are options loaded
-     */
-    if (!is_plugin_active('facebook-for-woocommerce/facebook-for-woocommerce.php') && count($options) > 0 && user_can(get_current_user_id(), 'manage_options')): ?>
+    </h2>
 
-        <form method="post">
+	<?php if ( 'delete-options' === $active_tab ) : ?>
 
-            <input type="hidden" name="reset" value="true">
-            <?php wp_nonce_field('deletefacebook_' . get_current_user_id()); ?>
+        <h3>Delete all WooCommerce for Facebook options</h3>
+        <h4>This is only for when you want to completely reset your connection with Facebook, after all else has
+            failed.</h4>
 
+        <h4>You have <?php echo esc_html( count( $options ) ); ?> Facebook for WooCommerce options stored in your
+            database.</h4>
 
-            <?php echo submit_button($text = "Delete all options", $type = 'delete button-primary', $name = 'delete_options', $wrap = true, array('style' => 'background: #d63638;border: #d63638;')); ?>
+		<?php
+		/**
+		 * Only if Facebook for WooCommerce is not active, and there are options loaded
+		 */
+		if ( ! is_plugin_active( 'facebook-for-woocommerce/facebook-for-woocommerce.php' ) && count( $options ) > 0 && user_can( get_current_user_id(), 'manage_options' ) ): ?>
 
-        </form>
+            <form method="post">
 
-    <?php endif; ?>
+                <input type="hidden" name="reset" value="true">
+				<?php wp_nonce_field( 'deletefacebook_' . get_current_user_id() ); ?>
 
-    <?php
-    /**
-     * Disable if Facebook for WooCommerce is active
-     */
-    if (is_plugin_active('facebook-for-woocommerce/facebook-for-woocommerce.php')): ?>
+				<?php
+                    echo submit_button(
+				        $text = "Delete all options",
+                        $type = 'delete button-primary',
+                        null,
+                        $wrap = true,
+                        array( 'style' => 'background: #d63638;border: #d63638;' )
+                    );
+				?>
 
-        <h3>Please deactivate WooCommerce for Facebook if you want to delete options.</h3>
+            </form>
 
-        <button class="button button-disabled">
-            Delete all options
-        </button>
+		<?php endif; ?>
 
-    <?php endif; ?>
+		<?php
+		/**
+		 * Disable if Facebook for WooCommerce is active
+		 */
+		if ( is_plugin_active( 'facebook-for-woocommerce/facebook-for-woocommerce.php' ) ): ?>
 
-    <br>
-    <br>
-    <hr>
+            <h3>Please deactivate WooCommerce for Facebook if you want to delete options.</h3>
 
-    <?php
-    /**
-     * Connection test requires Facebook for WooCommerce to be active
-     */
-    if (!is_plugin_active('facebook-for-woocommerce/facebook-for-woocommerce.php')): ?>
+            <button class="button button-disabled">
+                Delete all options
+            </button>
 
-        <h3>Run a connection test</h3>
+		<?php endif; ?>
 
+	<?php endif; ?>
 
-        <form method="post">
+	<?php if ( 'connection-test' === $active_tab ) : ?>
 
-            <input type="hidden" name="proxy" value="true">
-            <?php wp_nonce_field('proxyconnection_' . get_current_user_id()); ?>
+		<?php
+		/**
+		 * Connection test requires Facebook for WooCommerce to be active
+		 */
+		if ( ! is_plugin_active( 'facebook-for-woocommerce/facebook-for-woocommerce.php' ) ): ?>
 
+            <form method="post" action="options.php">
 
-            <fieldset>
-                <legend class="screen-reader-text"><span>Proxy Facebook Connectione</span></legend>
-                <label for="proxyconnection">
-                    <input name="proxyconnection" type="checkbox" id="proxyconnection"
-                           value="1" <?php checked(get_option('tfb4wc_set_proxy'), 1); ?> />
-                    <span>Enable the proxy connection test.</span>
-                </label>
-            </fieldset>
+	            <?php settings_fields( 'facebook_connection_test' ); ?>
+	            <?php do_settings_sections( 'troubleshoot-facebook-connection-test' ); ?>
+				<?php echo submit_button( 'Save proxy settings' ); ?>
 
-            <?php echo submit_button($text = "Save proxy settings", $type = 'primary', $name = 'submit', $wrap = true, $other_attributes = null); ?>
+            </form>
 
-        </form>
+		<?php endif; ?>
 
-
-    <?php endif; ?>
-
+	<?php endif; ?>
 
 </div>
